requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
PyJWT
